+++ Assets/Scripts/Creature/Boss/Component/SkillIndicator/CircleIndicator.cs
@@ -2,24 +2,27 @@ using Creature.Boss.Component;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-public class CircleIndicator : IndicatorBase
+namespace Creature.Boss.Component.SkillIndicator
 {
-    [SerializeField] protected GameObject circlePrefab;
-    public override void Initialize(Vector3 startPosition, Vector3 direction, float range, float duration)
+    public class CircleIndicator : IndicatorBase
     {
-        indicator = Instantiate(circlePrefab, startPosition, Quaternion.identity).transform;
-    }
-    public override IEnumerator GrowIndicator(float range, float duration)
-    {
-        float elapsedTime = 0f;
-        while (elapsedTime < duration)
+        [SerializeField] protected GameObject circlePrefab;
+        public override void Initialize(Vector3 startPosition, Vector3 direction, float range, float duration)
         {
-            float scaleX = Mathf.Lerp(0, range, elapsedTime / duration);
-            indicator.localScale = new Vector3(scaleX, scaleX, 1);
-            elapsedTime += Time.deltaTime;
-            yield return null;
+            indicator = Instantiate(circlePrefab, startPosition, Quaternion.identity).transform;
         }
-        yield return new WaitForSeconds(0.2f);
-        Destroy(indicator.gameObject);
+        public override IEnumerator GrowIndicator(float range, float duration)
+        {
+            float elapsedTime = 0f;
+            while (elapsedTime < duration)
+            {
+                float scaleX = Mathf.Lerp(0, range, elapsedTime / duration);
+                indicator.localScale = new Vector3(scaleX, scaleX, 1);
+                elapsedTime += Time.deltaTime;
+                yield return null;
+            }
+            yield return new WaitForSeconds(0.2f);
+            Destroy(indicator.gameObject);
+        }
     }
 }
